requirements.txt
requests>=2.31.0
pydantic>=2.6.0
google-generativeai>=0.7.2
rapidfuzz>=3.6.1
